Feat: 회원가입, 이메일/닉네임 중복체크 메서드

+++ backend/user/src/main/java/com/tosi/user/repository/ChildRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ChildRepository extends JpaRepository<Child, Integer> {
+public interface ChildRepository extends JpaRepository<Child, Long> {
 //    @Modifying
 //    @Query("delete from Child c where c.userId = :userId")
 //    void deleteByUserId(@Param("userId") Integer userId);
